Section BlueHeader renamed to MainForm
